fix: remove redundant cast in Hyprland monitor id comparison
hyprland-rs's MonitorId is already i128, so casting a Client's monitor
id to i128 before comparing against Monitor.id was a no-op flagged by
clippy.

+++ crates/services/src/compositor/hyprland.rs
@@ -29,7 +29,7 @@ pub fn is_available() -> bool {
 fn map_client(c: Client, monitors: &[super::types::Monitor]) -> super::types::Window {
     let monitor = c
         .monitor
-        .and_then(|id| monitors.iter().find(|m| m.id == id as i128))
+        .and_then(|id| monitors.iter().find(|m| m.id == id))
         .map(|m| m.name.clone())
         .unwrap_or_default();
